sclang: delete implicit casting in slot constructor
You should always know what type you are putting in the slot, therefore, you should be explicit. C++ is not explicit, but does offer a way to enforce this by deleting a generic template function.

Also, adds ctests that fail to compile, ensuring this behavior doesn't creep into the code base. This is a bit weird, but check our the pyrslot_*.cpp files and the cmake list for how that works. Ideally, these would be static_asserts with std::invokable but I couldn't get it to work.
