+++ testsuite/interpreter/pyrslot_enum.cpp
@@ -1,0 +1,3 @@
+#include "PyrSlot.h"
+enum { Bang };
+const auto a = PyrSlot::make(Bang);
