+++ testsuite/interpreter/pyrslot_int64.cpp
@@ -1,0 +1,2 @@
+#include "PyrSlot.h"
+const auto a = PyrSlot::make(std::int64_t {});
