+++ testsuite/interpreter/pyrslot_int8.cpp
@@ -1,0 +1,2 @@
+#include "PyrSlot.h"
+const auto a = PyrSlot::make(std::int8_t {});
