+++ testsuite/interpreter/pyrslot_ptr.cpp
@@ -1,0 +1,3 @@
+#include "PyrSlot.h"
+const int i { 10 };
+const auto a = PyrSlot::make(&i);
